softer queries with OR and rename to find
