+++ src/azure-cli/setup.py
@@ -54,6 +54,7 @@ DEPENDENCIES = [
     'azure-cli-core',
     'azure-cli-documentdb',
     'azure-cli-feedback',
+    'azure-cli-find',
     'azure-cli-iot',
     'azure-cli-keyvault',
     'azure-cli-network',
@@ -62,7 +63,6 @@ DEPENDENCIES = [
     'azure-cli-redis',
     'azure-cli-resource',
     'azure-cli-role',
-    'azure-cli-search',
     'azure-cli-sql',
     'azure-cli-storage',
     'azure-cli-vm'
