+++ src/command_modules/azure-cli-find/MANIFEST.in
