+++ src/command_modules/azure-cli-find/README.rst
@@ -1,0 +1,4 @@
+Microsoft Azure CLI 'find' Command Module
+=============================================
+
+Fulltext search and indexing module for Azure CLI
