+++ src/command_modules/azure-cli-find/azure/cli/command_modules/find/_help.py
@@ -3,12 +3,9 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-import azure.cli.command_modules.search._help  # pylint: disable=unused-import
+from azure.cli.core.help_files import helps
 
-
-def load_params(_):
-    import azure.cli.command_modules.search._params  # pylint: disable=redefined-outer-name
-
-
-def load_commands():
-    import azure.cli.command_modules.search.commands  # pylint: disable=redefined-outer-name
+helps['search'] = """
+    type: command
+    short-summary: Find Azure CLI commands based on a given query
+"""
