+++ src/command_modules/azure-cli-find/azure/cli/command_modules/find/commands.py
@@ -3,13 +3,6 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-from azure.cli.core.commands import CliArgumentType, register_cli_argument
+from azure.cli.core.commands import cli_command
 
-# pylint: disable=line-too-long
-
-query = CliArgumentType(
-    options_list=('--q', '-q'),
-    help='Query text to search for.'
-)
-
-register_cli_argument('search', 'criteria', query)
+cli_command(__name__, 'find', 'azure.cli.command_modules.find.custom#find')
