+++ src/command_modules/azure-cli-find/setup.cfg
