--- src/command_modules/azure-cli-search/README.rst
@@ -1,4 +1,0 @@
-Microsoft Azure CLI 'search' Command Module
-=============================================
-
-You know, for search
